.history/backend/requirements_20250711131745.txt
3fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
aio_pika==9.3.1